refactor(auth): make authentication code more clear
