+++ huaweicloud/config/config.go
@@ -2,58 +2,25 @@ package config
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math"
-	"net/http"
-	"os"
 	"sync"
 	"time"
 
 	"github.com/chnsz/golangsdk"
-	huaweisdk "github.com/chnsz/golangsdk/openstack"
 	"github.com/chnsz/golangsdk/openstack/identity/v3/domains"
 	"github.com/chnsz/golangsdk/openstack/identity/v3/projects"
 	"github.com/chnsz/golangsdk/openstack/identity/v3/users"
 	"github.com/chnsz/golangsdk/openstack/obs"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/logging"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/jmespath/go-jmespath"
-	"github.com/mitchellh/go-homedir"
-
-	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/helper/pathorcontents"
 )
 
 const (
 	obsLogFile         string = "./.obs-sdk.log"
 	obsLogFileSize10MB int64  = 1024 * 1024 * 10
-	securityKeyURL     string = "http://169.254.169.254/openstack/latest/securitykey"
-	keyExpiresDuration int64  = 600
 )
-
-// CLI Shared Config
-type SharedConfig struct {
-	Current  string    `json:"current"`
-	Profiles []Profile `json:"profiles"`
-}
-
-type Profile struct {
-	Name             string `json:"name"`
-	Mode             string `json:"mode"`
-	AccessKeyId      string `json:"accessKeyId"`
-	SecretAccessKey  string `json:"secretAccessKey"`
-	SecurityToken    string `json:"securityToken"`
-	Region           string `json:"region"`
-	ProjectId        string `json:"projectId"`
-	DomainId         string `json:"domainId"`
-	AgencyDomainId   string `json:"agencyDomainId"`
-	AgencyDomainName string `json:"agencyDomainName"`
-	AgencyName       string `json:"agencyName"`
-}
 
 type Config struct {
 	AccessKey           string
@@ -65,17 +32,11 @@ type Config struct {
 	DomainName          string
 	IdentityEndpoint    string
 	Insecure            bool
-	Password            string
 	Region              string
 	TenantID            string
 	TenantName          string
 	Token               string
 	SecurityToken       string
-	Username            string
-	UserID              string
-	AgencyName          string
-	AgencyDomainName    string
-	DelegatedProject    string
 	Cloud               string
 	MaxRetries          int
 	TerraformVersion    string
@@ -104,6 +65,14 @@ type Config struct {
 	// SecurityKeyLock is used to make the accessing of SecurityKeyExpiresAt serial,
 	// prevent sending duplicate query metadata api
 	SecurityKeyLock *sync.Mutex
+
+	// Legacy
+	Username         string
+	UserID           string
+	Password         string
+	AgencyName       string
+	AgencyDomainName string
+	DelegatedProject string
 }
 
 func (c *Config) LoadAndValidate() error {
@@ -111,37 +80,13 @@ func (c *Config) LoadAndValidate() error {
 		return fmt.Errorf("max_retries should be a positive value")
 	}
 
-	err := fmt.Errorf("Must config token or aksk or username password to be authorized")
-
-	if c.Token != "" {
-		err = buildClientByToken(c)
-
-	} else if c.AccessKey != "" && c.SecretKey != "" {
-		err = buildClientByAKSK(c)
-
-	} else if c.Password != "" {
-		if c.Username == "" && c.UserID == "" {
-			err = fmt.Errorf("\"password\": one of `user_name, user_id` must be specified")
-		} else {
-			err = buildClientByPassword(c)
-		}
-	} else if c.SharedConfigFile != "" {
-		err = buildClientByConfig(c)
-
-	} else {
-		err = getAuthConfigByMeta(c)
-		if err != nil {
-			return fmt.Errorf("Error fetching Auth credentials from ECS Metadata API, AkSk or ECS agency must be provided: %s", err)
-		}
-		log.Printf("[DEBUG] Successfully got metadata security key, which will expire at: %s", c.SecurityKeyExpiresAt)
-		err = buildClientByAKSK(c)
-	}
+	err := buildClient(c)
 	if err != nil {
 		return err
 	}
 
 	if c.Region == "" {
-		return fmt.Errorf("region should be provided.")
+		return fmt.Errorf("region should be provided")
 	}
 
 	if c.HwClient != nil && c.HwClient.ProjectID != "" {
@@ -174,54 +119,6 @@ func (c *Config) LoadAndValidate() error {
 	return nil
 }
 
-func (c *Config) reloadSecurityKey() error {
-	err := getAuthConfigByMeta(c)
-	if err != nil {
-		return fmt.Errorf("Error reloading Auth credentials from ECS Metadata API: %s", err)
-	}
-	log.Printf("Successfully reload metadata security key, which will expire at: %s", c.SecurityKeyExpiresAt)
-	return buildClientByAKSK(c)
-}
-
-func generateTLSConfig(c *Config) (*tls.Config, error) {
-	config := &tls.Config{}
-	if c.CACertFile != "" {
-		caCert, _, err := pathorcontents.Read(c.CACertFile)
-		if err != nil {
-			return nil, fmt.Errorf("Error reading CA Cert: %s", err)
-		}
-
-		caCertPool := x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM([]byte(caCert))
-		config.RootCAs = caCertPool
-	}
-
-	if c.Insecure {
-		config.InsecureSkipVerify = true
-	}
-
-	if c.ClientCertFile != "" && c.ClientKeyFile != "" {
-		clientCert, _, err := pathorcontents.Read(c.ClientCertFile)
-		if err != nil {
-			return nil, fmt.Errorf("Error reading Client Cert: %s", err)
-		}
-		clientKey, _, err := pathorcontents.Read(c.ClientKeyFile)
-		if err != nil {
-			return nil, fmt.Errorf("Error reading Client Key: %s", err)
-		}
-
-		cert, err := tls.X509KeyPair([]byte(clientCert), []byte(clientKey))
-		if err != nil {
-			return nil, err
-		}
-
-		config.Certificates = []tls.Certificate{cert}
-		config.BuildNameToCertificate()
-	}
-
-	return config, nil
-}
-
 func retryBackoffFunc(ctx context.Context, respErr *golangsdk.ErrUnexpectedResponseCode, e error, retries uint) error {
 	minutes := int(math.Pow(2, float64(retries)))
 	if minutes > 30 { // won't wait more than 30 minutes
@@ -241,309 +138,6 @@ func retryBackoffFunc(ctx context.Context, respErr *golangsdk.ErrUnexpectedRespo
 		//lintignore:R018
 		time.Sleep(sleep)
 	}
-
-	return nil
-}
-
-func genClient(c *Config, ao golangsdk.AuthOptionsProvider) (*golangsdk.ProviderClient, error) {
-	client, err := huaweisdk.NewClient(ao.GetIdentityEndpoint())
-	if err != nil {
-		return nil, err
-	}
-
-	// Set UserAgent
-	client.UserAgent.Prepend("terraform-provider-iac")
-
-	config, err := generateTLSConfig(c)
-	if err != nil {
-		return nil, err
-	}
-	transport := &http.Transport{
-		Proxy:           http.ProxyFromEnvironment,
-		TLSClientConfig: config,
-	}
-
-	client.HTTPClient = http.Client{
-		Transport: &LogRoundTripper{
-			Rt:         transport,
-			OsDebug:    logging.IsDebugOrHigher(),
-			MaxRetries: c.MaxRetries,
-		},
-		CheckRedirect: func(req *http.Request, via []*http.Request) error {
-			if client.AKSKAuthOptions.AccessKey != "" {
-				golangsdk.ReSign(req, golangsdk.SignOptions{
-					AccessKey:  client.AKSKAuthOptions.AccessKey,
-					SecretKey:  client.AKSKAuthOptions.SecretKey,
-					RegionName: client.AKSKAuthOptions.Region,
-				})
-			}
-			return nil
-		},
-	}
-
-	if c.MaxRetries > 0 {
-		client.MaxBackoffRetries = uint(c.MaxRetries)
-		client.RetryBackoffFunc = retryBackoffFunc
-	}
-
-	// Validate authentication normally.
-	err = huaweisdk.Authenticate(client, ao)
-	if err != nil {
-		return nil, err
-	}
-
-	return client, nil
-}
-
-func buildClientByToken(c *Config) error {
-	var pao, dao golangsdk.AuthOptions
-
-	if c.AgencyDomainName != "" && c.AgencyName != "" {
-		pao = golangsdk.AuthOptions{
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-			DelegatedProject: c.DelegatedProject,
-		}
-
-		dao = golangsdk.AuthOptions{
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-		}
-	} else {
-		pao = golangsdk.AuthOptions{
-			DomainID:   c.DomainID,
-			DomainName: c.DomainName,
-			TenantID:   c.TenantID,
-			TenantName: c.TenantName,
-		}
-
-		dao = golangsdk.AuthOptions{
-			DomainID:   c.DomainID,
-			DomainName: c.DomainName,
-		}
-	}
-
-	for _, ao := range []*golangsdk.AuthOptions{&pao, &dao} {
-		ao.IdentityEndpoint = c.IdentityEndpoint
-		ao.TokenID = c.Token
-
-	}
-	return genClients(c, pao, dao)
-}
-
-func buildClientByAKSK(c *Config) error {
-	var pao, dao golangsdk.AKSKAuthOptions
-
-	if c.AgencyDomainName != "" && c.AgencyName != "" {
-		pao = golangsdk.AKSKAuthOptions{
-			DomainID:         c.DomainID,
-			Domain:           c.DomainName,
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-			DelegatedProject: c.DelegatedProject,
-		}
-
-		dao = golangsdk.AKSKAuthOptions{
-			DomainID:         c.DomainID,
-			Domain:           c.DomainName,
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-		}
-	} else {
-		pao = golangsdk.AKSKAuthOptions{
-			BssDomainID: c.DomainID,
-			BssDomain:   c.DomainName,
-			ProjectName: c.TenantName,
-			ProjectId:   c.TenantID,
-		}
-
-		dao = golangsdk.AKSKAuthOptions{
-			DomainID: c.DomainID,
-			Domain:   c.DomainName,
-		}
-	}
-
-	for _, ao := range []*golangsdk.AKSKAuthOptions{&pao, &dao} {
-		ao.IdentityEndpoint = c.IdentityEndpoint
-		ao.AccessKey = c.AccessKey
-		ao.SecretKey = c.SecretKey
-		if c.Region != "" {
-			ao.Region = c.Region
-		}
-		if c.SecurityToken != "" {
-			ao.SecurityToken = c.SecurityToken
-			ao.WithUserCatalog = true
-		}
-	}
-	return genClients(c, pao, dao)
-}
-
-func buildClientByConfig(c *Config) error {
-	profilePath, err := homedir.Expand(c.SharedConfigFile)
-	if err != nil {
-		return err
-	}
-
-	current := c.Profile
-	var providerConfig Profile
-	_, err = os.Stat(profilePath)
-	if !os.IsNotExist(err) {
-		data, err := ioutil.ReadFile(profilePath)
-		if err != nil {
-			return fmt.Errorf("Err reading from shared config file: %s", err)
-		}
-		sharedConfig := SharedConfig{}
-		err = json.Unmarshal(data, &sharedConfig)
-		if err != nil {
-			return err
-		}
-
-		// fetch current from shared config if not specified with provider
-		if current == "" {
-			current = sharedConfig.Current
-		}
-
-		// fetch the current profile config
-		for _, v := range sharedConfig.Profiles {
-			if current == v.Name {
-				providerConfig = v
-				break
-			}
-		}
-		if (providerConfig == Profile{}) {
-			return fmt.Errorf("Error finding profile %s from shared config file", current)
-		}
-	} else {
-		return fmt.Errorf("The specified shared config file %s does not exist", profilePath)
-	}
-
-	if providerConfig.Mode == "AKSK" {
-		c.AccessKey = providerConfig.AccessKeyId
-		c.SecretKey = providerConfig.SecretAccessKey
-		if providerConfig.Region != "" {
-			c.Region = providerConfig.Region
-		}
-		// non required fields
-		if providerConfig.DomainId != "" {
-			c.DomainID = providerConfig.DomainId
-		}
-		if providerConfig.ProjectId != "" {
-			c.TenantID = providerConfig.ProjectId
-		}
-	} else {
-		return fmt.Errorf("Unsupported mode %s in shared config file", providerConfig.Mode)
-	}
-	return buildClientByAKSK(c)
-}
-
-func buildClientByPassword(c *Config) error {
-	var pao, dao golangsdk.AuthOptions
-
-	if c.AgencyDomainName != "" && c.AgencyName != "" {
-		pao = golangsdk.AuthOptions{
-			DomainID:         c.DomainID,
-			DomainName:       c.DomainName,
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-			DelegatedProject: c.DelegatedProject,
-		}
-
-		dao = golangsdk.AuthOptions{
-			DomainID:         c.DomainID,
-			DomainName:       c.DomainName,
-			AgencyName:       c.AgencyName,
-			AgencyDomainName: c.AgencyDomainName,
-		}
-	} else {
-		pao = golangsdk.AuthOptions{
-			DomainID:   c.DomainID,
-			DomainName: c.DomainName,
-			TenantID:   c.TenantID,
-			TenantName: c.TenantName,
-		}
-
-		dao = golangsdk.AuthOptions{
-			DomainID:   c.DomainID,
-			DomainName: c.DomainName,
-		}
-	}
-
-	for _, ao := range []*golangsdk.AuthOptions{&pao, &dao} {
-		ao.IdentityEndpoint = c.IdentityEndpoint
-		ao.Password = c.Password
-		ao.Username = c.Username
-		ao.UserID = c.UserID
-	}
-	return genClients(c, pao, dao)
-}
-
-func genClients(c *Config, pao, dao golangsdk.AuthOptionsProvider) error {
-	client, err := genClient(c, pao)
-	if err != nil {
-		return err
-	}
-	c.HwClient = client
-
-	client, err = genClient(c, dao)
-	if err == nil {
-		c.DomainClient = client
-	}
-	return err
-}
-
-func getAuthConfigByMeta(c *Config) error {
-	req, err := http.NewRequest("GET", securityKeyURL, nil)
-	if err != nil {
-		return fmt.Errorf("Error building metadata API request: %s", err.Error())
-	}
-
-	httpClient := &http.Client{}
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return fmt.Errorf("Error requesting metadata API: %s", err.Error())
-	}
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("Error requesting metadata API: status code = %d", resp.StatusCode)
-	}
-
-	var parsedBody interface{}
-
-	defer resp.Body.Close()
-	rawBody, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("Error parsing metadata API response: %s", err.Error())
-	}
-
-	err = json.Unmarshal(rawBody, &parsedBody)
-	if err != nil {
-		return fmt.Errorf("Error unmarshal metadata API, agency_name is empty: %s", err.Error())
-	}
-
-	expiresAt, err := jmespath.Search("credential.expires_at", parsedBody)
-	if err != nil {
-		return fmt.Errorf("Error fetching metadata expires_at: %s", err.Error())
-	}
-	accessKey, err := jmespath.Search("credential.access", parsedBody)
-	if err != nil {
-		return fmt.Errorf("Error fetching metadata access: %s", err.Error())
-	}
-	secretKey, err := jmespath.Search("credential.secret", parsedBody)
-	if err != nil {
-		return fmt.Errorf("Error fetching metadata secret: %s", err.Error())
-	}
-	securityToken, err := jmespath.Search("credential.securitytoken", parsedBody)
-	if err != nil {
-		return fmt.Errorf("Error fetching metadata securitytoken: %s", err.Error())
-	}
-
-	if accessKey == nil || secretKey == nil || securityToken == nil || expiresAt == nil {
-		return fmt.Errorf("Error fetching metadata authentication information.")
-	}
-	expairesTime, err := time.Parse(time.RFC3339, expiresAt.(string))
-	if err != nil {
-		return err
-	}
-	c.AccessKey, c.SecretKey, c.SecurityToken, c.SecurityKeyExpiresAt = accessKey.(string), secretKey.(string), securityToken.(string), expairesTime
 
 	return nil
 }
